src/test/java/dev/dobicinaitis/advent/Day6Test.java
package dev.dobicinaitis.advent;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day6Test {

    private Puzzle puzzle = new Day6();

    @Test
    public void firstPuzzleExampleTest() {
        assertEquals(5934, puzzle.getFirstPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    public void firstPuzzleMyInputTest() {
        assertEquals(389726, puzzle.getFirstPuzzleSolution(puzzle.myInputFilename));
    }

    @Test
    public void secondPuzzleExampleTest() {
        assertEquals(26984457539L, puzzle.getSecondPuzzleSolution(puzzle.exampleInputFilename));
    }

    @Test
    public void secondPuzzleMyInputTest() {
        assertEquals(1743335992042L, puzzle.getSecondPuzzleSolution(puzzle.myInputFilename));
    }
}